Glasses/requirements.txt
asgiref==3.5.1
Django==3.2
django-environ==0.8.1
django-filter==21.1
django-fsm==2.8.0
django-model-utils==4.2.0
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
djangorestframework-recursive==0.1.2
drf-nested-routers==0.93.4
drf-secure-token==1.1.0
future==0.18.2
mysqlclient==2.1.0
phonenumbers==8.12.48
Pillow==9.1.0
pytz==2022.1
serializers==0.2.4
sqlparse==0.4.2
unicef-djangolib==0.5.4
unicef-restlib==0.7
uuid==1.30

